Events changes
Changed function signature for pushEvent to use a raw ptr to not have to type make_unique every time to make an event and all that. To not have the run out of scope issue, when calling execute() on the object after it goes out of scope, I
added clone function to clone derived objects easily. To not add too much boiler plate, I took from Cherno and used macro expansion to do MODEL_EVENT_CLONE(type name) to simply define the same function over and over. I did the same with UI event naturally.